requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
openai==1.3.0
python-dotenv==0.19.0
autogen==1.0.16
pyautogen==0.1.14
requests==2.31.0 